+++ CHANGELOG.md
@@ -12,6 +12,14 @@ Add changes in new features here. Do not change the gem's version in pull/merge 
 ### Changes
 -
 
+## [0.2.0] - 2025-08-13
+
+[Diff](https://github.com/Verseth/ruby-nested_hash_builder/compare/v0.1.0...v0.2.0)
+
+### Changes
+- Added `NestedHashBuilder.build!` for strict typechecking
+- Added `NestedHashBuilder::Proxy#[]=` and `NestedHashBuilder::Proxy#[]`
+
 ## [0.1.0] - 2025-08-13
 
 [Diff](https://github.com/Verseth/ruby-nested_hash_builder/compare/v0.0.0...v0.1.0)